ocr-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pymupdf==1.23.6
easyocr==1.7.1
uuid==1.30
python-dotenv==1.0.0